cursor movement almost implemented, just have to connect everything

+++ app/src/main/java/Game/Utility/Point.java
@@ -5,17 +5,17 @@ package Game.Utility;
  */
 public class Point {
 
-    public int x, y;
+    public float x, y;
 
-    public Point (int x, int y){
+    public Point (float x, float y){
         this.x = x;
         this.y = y;
     }
-    public int getX(){
+    public float getX(){
         return x;
     }
 
-    public int getY(){
+    public float getY(){
         return y;
     }
 }
